Skip seeders on production

+++ src/seeders/20200503163557-add-careers.ts
@@ -1,8 +1,11 @@
 import { QueryInterface } from "sequelize";
 import { careerCodes } from "./constants/careerCodes";
+import { Environment } from "../config/Environment";
 
 export = {
-  up: (queryInterface: QueryInterface) => {
+  up: async (queryInterface: QueryInterface) => {
+    if (Environment.NODE_ENV() === Environment.PRODUCTION) return;
+
     return queryInterface.bulkInsert("Careers", [
       {
         code: careerCodes.IngenieriaCivil,
@@ -84,7 +87,8 @@ export = {
       }
     ]);
   },
-  down: (queryInterface: QueryInterface) => {
+  down: async (queryInterface: QueryInterface) => {
+    if (Environment.NODE_ENV() === Environment.PRODUCTION) return;
     return queryInterface.bulkDelete("Careers", {}, {});
   }
 };
