+++ src/seeders/20200612002940-add-admins.ts
@@ -1,11 +1,14 @@
 import { QueryInterface } from "sequelize";
 import { federico, aldana } from "./constants/admins";
+import { Environment } from "../config/Environment";
 
 export = {
-  up: (queryInterface: QueryInterface) => {
+  up: async (queryInterface: QueryInterface) => {
+    if (Environment.NODE_ENV() === Environment.PRODUCTION) return;
     return queryInterface.bulkInsert("Admins", [federico.admin, aldana.admin]);
   },
-  down: (queryInterface: QueryInterface) => {
+  down: async (queryInterface: QueryInterface) => {
+    if (Environment.NODE_ENV() === Environment.PRODUCTION) return;
     return queryInterface.bulkDelete("Admins", {});
   }
 };
